added srcdir to node path
