+++ src/node_path/src/GCD.js
