+++ src/node_path/src/Intent.js
